requirements.txt
einops==0.8.0
imageio==2.34.2
numpy==1.26.4
transformers==4.39.3
open_clip_torch==2.22.0
Pillow==10.4.0
tqdm==4.66.4
PyYAML==6.0.1
setuptools==69.5.1
scikit-learn==1.5.1
lightning==2.4.0
deepspeed==0.14.5
open3d==0.18.0
decord
omegaconf
opencv_python
pandas
moviepy
av
gradio
timm
kornia
fairscale
diffusers